feat: Made task delete working properly

+++ frontend/src/components/task/TaskItems.js
@@ -1,7 +1,9 @@
 import React from "react";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
-import { deleteTaskAsync } from "../../redux/slices/taskSlice";
+import { deleteTaskAsync, fetchTasks } from "../../redux/slices/taskSlice";
+import { toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 import "../../styles/TaskItem.css";
 
 const TaskItem = ({ task }) => {
@@ -12,8 +14,23 @@ const TaskItem = ({ task }) => {
     navigate(`/edit-task/${task._id}`);
   };
 
-  const handleDelete = () => {
-    dispatch(deleteTaskAsync(task._id));
+  const handleDelete = async () => {
+    const confirmed = window.confirm(
+      "Are you sure you want to delete this task?"
+    );
+    if (confirmed) {
+      try {
+        await dispatch(deleteTaskAsync(task._id));
+        toast.success("Task deleted successfully", {
+          position: "top-right",
+        });
+        dispatch(fetchTasks());
+      } catch (error) {
+        toast.error("Failed to delete task", {
+          position: "top-right",
+        });
+      }
+    }
   };
 
   return (
